Edit dashboard.

+++ src/main/resources/db/migration/V1__create_securities_table.sql
@@ -2,8 +2,8 @@ CREATE TABLE securities
 (
     id                          UUID PRIMARY KEY,
     ISIN                        VARCHAR(12),
-    min_amount_transaction      INTEGER,
-    rating                      INTEGER,
-    profitability_for_last_year INTEGER,
+    min_amount_transaction      BIGINT,
+    rating                      BIGINT,
+    profitability_for_last_year BIGINT,
     additional_info             VARCHAR(255)
 );
